requirements.txt
a2a-sdk>=0.2.5
click>=8.1.8
dotenv>=0.9.9
httpx>=0.28.1
langchain-core>=0.3.0
langchain-google-genai>=2.1.4
langchain-openai>=0.2.9
langgraph>=0.4.1
pydantic>=2.11.4
python-dotenv>=1.1.0
uvicorn>=0.34.2